Refactor data class file structure
