+++ app-mobile-tracker/src/main/java/kaist/iclab/mobiletracker/data/campaign/CampaignSensorList.kt
@@ -1,4 +1,4 @@
-package kaist.iclab.mobiletracker.data.sensors
+package kaist.iclab.mobiletracker.data.campaign
 
 import kotlinx.serialization.Serializable
