+++ app-mobile-tracker/src/main/java/kaist/iclab/mobiletracker/data/campaign/ProfileData.kt
@@ -1,4 +1,4 @@
-package kaist.iclab.mobiletracker.data.sensors
+package kaist.iclab.mobiletracker.data.campaign
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -17,4 +17,3 @@ data class ProfileData(
     val campaignId: Int? = null,
     val email: String = ""
 )
-
