+++ app-mobile-tracker/src/main/java/kaist/iclab/mobiletracker/repository/UserProfileRepository.kt
@@ -1,6 +1,6 @@
 package kaist.iclab.mobiletracker.repository
 
-import kaist.iclab.mobiletracker.data.sensors.ProfileData
+import kaist.iclab.mobiletracker.data.campaign.ProfileData
 import kotlinx.coroutines.flow.StateFlow
 
 /**
