+++ app-mobile-tracker/src/main/java/kaist/iclab/mobiletracker/storage/UserProfileStorage.kt
@@ -1,6 +1,6 @@
 package kaist.iclab.mobiletracker.storage
 
-import kaist.iclab.mobiletracker.data.sensors.ProfileData
+import kaist.iclab.mobiletracker.data.campaign.ProfileData
 import kaist.iclab.tracker.storage.couchbase.CouchbaseDB
 import kaist.iclab.tracker.storage.couchbase.CouchbaseStateStorage
