+++ app-mobile-tracker/src/main/java/kaist/iclab/mobiletracker/viewmodels/settings/AccountSettingsViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kaist.iclab.mobiletracker.R
 import kaist.iclab.mobiletracker.data.campaign.CampaignData
-import kaist.iclab.mobiletracker.data.sensors.CampaignTableData
+import kaist.iclab.mobiletracker.data.campaign.CampaignTableData
 import kaist.iclab.mobiletracker.repository.AppError
 import kaist.iclab.mobiletracker.repository.CampaignRepository
 import kaist.iclab.mobiletracker.repository.CampaignSensorRepository
